Major refactor, focus on multicamera/best OOP practices
