+++ GTAV_CustomCamera.csproj
@@ -53,6 +53,7 @@
     <Compile Include="Camera\CustomCamera.cs" />
     <Compile Include="Camera\LegacyCamera.cs" />
     <Compile Include="Camera\SmoothCamera.cs" />
+    <Compile Include="Camera\ThirdPersonCamera.cs" />
     <Compile Include="Helper\InputR.cs" />
     <Compile Include="Helper\MathR.cs" />
     <Compile Include="Helper\Time.cs" />
